unsignedInteger to unsignedBigInteger

+++ WECOP/database/migrations/2021_03_21_000000_create_orders_table.php
@@ -19,9 +19,9 @@ class CreateOrdersTable extends Migration
             $table->text('payment_type');
             $table->timestamp('date')->useCurrent();
             $table->double('total');
-            $table->unsignedInteger('address');
+            $table->unsignedBigInteger('address');
             $table->foreign('address')->references('id')->on('addresses'); 
-            $table->unsignedInteger('user');
+            $table->unsignedBigInteger('user');
             $table->foreign('user')->references('id')->on('users'); 
         });
     }
